Implement user-to-user WebRTC DataChannels
Users can click on call next to another user's name to initiate
a direct connection to that user. Currently the caller will open
`this.sendDataChannel` and the receiever will open
`this.receieveDataChannel`
This should be fixable at app.js line 92ish and 145.
Basically when we get the remote data channel we need to do things
a bit differently.

Also functionality for actually sending data back and forth on the
channel needs to be implemented.